requirements.txt
streamlit
SpeechRecognition
gtts
googletrans==4.0.0rc1
google-generativeai
